requirements.txt
Flask==0.12.2
gunicorn==19.6.0
line-bot-sdk==1.3.0
python-dotenv==0.8.2
SQLAlchemy==1.2.7